WIP audit tests
Work in progress tests for the audits page. Currently tests correctly for logging in as an admin and verifying that they can see and access the audit page. Currently working on testing the date filter functionality for the audit page.

Still need to test as other users to make sure they are not able to see the page.

Also contains a useful method for logging in as a specified user type